Use IActtionResults to provide a more RESTful way of responding to requests.

+++ TransactionWebApi/TransactionWebApi/Controllers/TransactionsController.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using TransactionWebApi.Models;
 using TransactionWebApi.Repositories;
 
@@ -16,32 +12,59 @@ namespace TransactionWebApi.Controllers {
 		}
 		// GET api/transactions
 		[HttpGet]
-		public IEnumerable<Transaction> Get() {
-			return _transactionRepository.GetAll();
+		public IActionResult Get() {
+			return Ok(_transactionRepository.GetAll());
 		}
 
 		// GET api/transactions/5
 		[HttpGet("{id}")]
-		public Transaction Get(int id) {
-			return _transactionRepository.Get(id);
+		public IActionResult Get(int id) {
+			Transaction foundTransaction = _transactionRepository.Get(id);
+			if (foundTransaction == null) {
+				return NotFound(id);
+			}
+
+			return Ok(foundTransaction);
 		}
 
 		// POST api/transactions
 		[HttpPost]
-		public void Post([FromBody]Transaction transaction) {
-			_transactionRepository.Add(transaction);
+		public IActionResult Post([FromBody]Transaction transaction) {
+			if (transaction == null) {
+				return BadRequest(transaction);
+			}
+			Transaction addedTransaction = _transactionRepository.Add(transaction);
+
+			// Return the created object included generated ID and Date fields.
+			return Ok(addedTransaction);
 		}
 
 		// PUT api/transactions/5
 		[HttpPut("{id}")]
-		public void Put(int id, [FromBody]Transaction transaction) {
+		public IActionResult Put(int id, [FromBody]Transaction transaction) {
+			
+			if (transaction == null) {
+				return BadRequest(transaction);
+			}
+
+			Transaction transactionToUpdate = _transactionRepository.Get(id);
+			if (transactionToUpdate == null) {
+				return NotFound(transaction);
+			}
+
 			_transactionRepository.Update(id, transaction);
+			return Ok();
 		}
 
 		// DELETE api/transactions/5
 		[HttpDelete("{id}")]
-		public void Delete(int id) {
+		public IActionResult Delete(int id) {
+			Transaction transactionToDelete = _transactionRepository.Get(id);
+			if (transactionToDelete == null) {
+				return NotFound(id);
+			}
 			_transactionRepository.Delete(id);
+			return Ok();
 		}
 	}
 }
